setRow

+++ src/main/java/controller/Main.java
@@ -5,6 +5,7 @@ import model.dto.Product;
 import util.Validate;
 import view.View;
 
+import java.io.*;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Scanner;
@@ -14,19 +15,23 @@ public class Main {
     public static void main(String[] args) {
         ProductImplement model = new ProductImplement();
         int start = 0;
-        int limite = 2;
+        int limite;
+        try {
+            BufferedReader reader = new BufferedReader(new FileReader("D:\\HRD\\Java_HRD\\Mini\\mini_project\\src\\main\\java\\util\\setRow.txt"));
+            int data = Integer.parseInt(reader.readLine());
+            limite = data;
+            reader.close();
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
         int cur_page = 1;
         int totalItems = model.count();
         int totalPages = (int) Math.ceil((double) totalItems / limite);
-        do {
+        back : do {
             View view = new View();
             int count = model.count();
-            try {
-                ResultSet rs = model.selectAll(limite,start);
-                view.display(rs,limite,count,cur_page);
-            }catch (SQLException ex){
-                System.out.println(ex.getMessage());
-            }
+            ResultSet rs = model.selectAll(limite,start);
+            view.display(rs,limite,count,cur_page);
             System.out.println("F) First\t\t P) Previous\t\t N) Next\t\t L) Last\t\t G) Go to");
             System.out.println();
             System.out.println("================================================== Display ==================================================");
@@ -35,22 +40,11 @@ public class Main {
             System.out.println("=============================================================================================================");
             System.out.print("Choose Option : ");
             String option = scanner.next();
-//            try {
-//                if(!Validate.validate_option_char(option)){
-//                    do {
-//                        System.out.println("Option do not match !!!!");
-//                        System.out.print("Choose Option : ");
-//                        option = scanner.next();
-//                    }while (!Validate.validate_option_char(option));
-//                }
-//            }catch (Exception e){
-//                System.out.println(e.getMessage());
-//            }
             switch (option){
                 case "F" ->{
                     cur_page=1;
                     start = 0;
-                    break;
+                    continue back;
                 }
                 case "P" ->{
                     if(cur_page > 1){
@@ -58,32 +52,34 @@ public class Main {
                     }
                     if(start != 0){
                         start = start-limite;
+                        continue back;
                     }else{
                         start = 0;
+                        continue back;
                     }
-                    break;
                 }
                 case "N" ->{
                     if(cur_page != totalPages){
                         cur_page++;
-                    }else{
-                        start = totalItems - limite;
                     }
-                    if(start < totalItems-limite){
+                    if(start <totalItems-limite){
                         start = start+limite;
+                        continue back;
                     }else{
-                        start = (totalItems - 1) * limite;
+                        start = (totalPages - 1) * limite;
+                        continue back;
                     }
-                    break;
                 }
                 case "L" ->{
                     cur_page = totalPages;
+                    int last = totalItems%limite;
                     if(limite == 2){
-                        start = totalItems-(totalItems % limite) - 2 ;
+                        start = totalItems-(totalItems%limite) - 2 ;
+                        continue back;
                     }else{
-                        start = totalItems-(totalItems % limite);
+                        start = totalItems-(totalItems%limite);
+                        continue back;
                     }
-                    break;
                 }
                 case "G" ->{
                     System.out.print("Enter page :");
@@ -92,11 +88,11 @@ public class Main {
                     if(page_re <= totalPages){
                         cur_page = page_re;
                         start = (page_re - 1) * limite;
+                        continue back;
                     }else{
                         System.out.println("Out off limite page");
-                        return;
+                        break;
                     }
-                    break;
                 }
                 case "W" ->{
                     Product insert = view.write();
@@ -107,30 +103,66 @@ public class Main {
                 case "R" ->{
                     System.out.print("Enter ID to show product : ");
                     String id = scanner.next();
-                    ResultSet rs = model.view(id);
-                    view.read(rs);
+                    ResultSet rs_read = model.view(id);
+                    view.read(rs_read);
+                    System.out.print("Press enter for continues....");
+                    scanner.nextLine();
+                    scanner.nextLine();
                     break ;
                 }
                 case "U" ->{
-
+                    System.out.print("Enter ID to update product : ");
+                    String id = scanner.next();
+                    Product rs_update = view.update();
+                    model.update(Integer.parseInt(id),rs_update);
+                    System.out.println("Product update successfully !!!");
                     break;
                 }
                 case "D" ->{
                     System.out.print("Enter ID to delete product : ");
                     String id = scanner.next();
-                    ResultSet rs = model.delete(id);
-                    view.delete(rs);
+                    ResultSet rs_delete = model.view(id);
+                    view.read(rs_delete);
+                    System.out.print("Enter Y for confirm or B for back to display :");
+                    String ans = scanner.nextLine();
+                    scanner.nextLine();
+                    switch (ans){
+                        case "y" ->{
+                            model.delete(id);
+                        }
+                        case "b" ->{
+                            continue back;
+                        }
+                    }
                     break;
                 }
                 case "S" ->{
                     System.out.print("Enter product name to search : ");
                     String search = scanner.next();
-                    ResultSet rs = model.search(search);
-                    view.search(rs);
+                    ResultSet rs_search = model.search(search);
+                    view.search(rs_search);
                     break;
                 }
                 case "Se" ->{
-                    break;
+                    System.out.print("Enter number of Row :");
+                    String num_row = scanner.next();
+                    scanner.nextLine();
+                    try{
+                        BufferedWriter writer = new BufferedWriter(new FileWriter("D:\\HRD\\Java_HRD\\Mini\\mini_project\\src\\main\\java\\util\\setRow.txt"));
+                        writer.write(num_row);
+                        writer.close();
+                    }catch(IOException e){
+                        e.printStackTrace();
+                    }
+                    try {
+                        BufferedReader reader = new BufferedReader(new FileReader("D:\\HRD\\Java_HRD\\Mini\\mini_project\\src\\main\\java\\util\\setRow.txt"));
+                        int data = Integer.parseInt(reader.readLine());
+                        limite = data;
+                        reader.close();
+                    } catch (Exception e) {
+                        throw new RuntimeException(e);
+                    }
+                    continue back;
                 }
                 case "Sa" ->{
                     break;
